fix(vesting-routes): save & revert moved to utils
